% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{get_mappings}
\alias{get_mappings}
\title{Get Likert Mappings}
\usage{
get_mappings(custom_mappings = NULL)
}
\arguments{
\item{custom_mappings}{A list of custom mappings. If NULL, the default mappings are loaded.}
}
\value{
A list containing the Likert scale mappings.
}
\description{
This function loads either the default or custom Likert scale mappings.
}
