% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_mappings}
\alias{load_mappings}
\title{Load Likert Mappings
This function loads Likert scale mappings from a specified JSON file.}
\usage{
load_mappings(file_path = NULL)
}
\arguments{
\item{file_path}{The path to the JSON file containing the mappings.}
}
\value{
A list containing the Likert scale mappings.
}
\description{
Load Likert Mappings
This function loads Likert scale mappings from a specified JSON file.
}
