% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tranform.R
\name{transform_column}
\alias{transform_column}
\title{Transform a Single Column}
\usage{
transform_column(column, mapping, unmapped_action = NA)
}
\arguments{
\item{column}{The column to be transformed.}

\item{mapping}{The mapping to be used for transformation.}

\item{unmapped_action}{The action to take for unmapped values.}
}
\value{
A transformed column.
}
\description{
This function transforms a single column based on a given mapping.
}
