% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tranform.R
\name{transform_likert}
\alias{transform_likert}
\title{Transform Likert Columns}
\usage{
transform_likert(
  data,
  variables,
  mapping_name,
  custom_mappings = NULL,
  unmapped_action = NA
)
}
\arguments{
\item{data}{The dataframe containing the columns to be transformed.}

\item{variables}{The columns to be transformed.}

\item{mapping_name}{The name of the mapping to be used.}

\item{custom_mappings}{A list of custom mappings. If NULL, the default mappings are loaded.}

\item{unmapped_action}{The action to take for unmapped values.}
}
\value{
A dataframe with transformed columns.
}
\description{
This function transforms multiple columns using the same Likert scale mapping.
}
