% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tranform.R
\name{transform_likert_multiple}
\alias{transform_likert_multiple}
\title{Transform Multiple Likert Columns}
\usage{
transform_likert_multiple(
  data,
  variable_mapping_list,
  custom_mappings = NULL,
  unmapped_action = NA
)
}
\arguments{
\item{data}{The dataframe containing the columns to be transformed.}

\item{variable_mapping_list}{A list mapping variable names to their respective mappings.}

\item{custom_mappings}{A list of custom mappings. If NULL, the default mappings are loaded.}

\item{unmapped_action}{The action to take for unmapped values.}
}
\value{
A dataframe with transformed columns.
}
\description{
This function transforms multiple columns using different Likert scale mappings.
}
